backend/requirements.txt
fastapi
uvicorn[standard]
transformers
sentence-transformers
torch
faiss-cpu
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
tqdm